requirements.txt
azure-functions
requests
msal
beautifulsoup4
lxml
